+++ client/src/app/Pixi/GameBoard/GridObject.ts
