+++ client/src/app/Pixi/GameBoard/TurnLabel.ts
@@ -1,6 +1,6 @@
 import * as PIXI from 'pixi.js';
 import { PixiManager } from '../../../api/PixiManager';
-import { CHAR_H } from '../../Pixi/Utils/FontLoader';
+import { CHAR_H } from '../Utils/FontLoader';
 import { PixiObject } from '../PixiObject';
 import { TextAlign, TextObject } from '../Text';
 import { GameGrid } from './GameGrid';
