+++ client/src/app/Pixi/ObjectiveManager.ts
