+++ client/src/app/Pixi/ObjectiveObject.ts
