+++ client/src/app/Pixi/ObjectiveSprite.ts
