+++ client/src/app/Pixi/PieceObject.ts
