+++ client/src/app/Pixi/PixiObject.ts
