+++ client/src/app/Pixi/PixiUtils.ts
