+++ client/src/app/Pixi/ResourceBars.ts
