+++ client/src/app/Pixi/ShipManager.ts
