+++ client/src/app/Pixi/ShipSprite.ts
