+++ client/src/app/Pixi/ShopItem.ts
