+++ client/src/app/Pixi/ShopModal.ts
