+++ client/src/app/Pixi/StatIcon.ts
