+++ client/src/app/Pixi/Text.ts
