+++ client/src/app/Pixi/TextureLoader.ts
